+++ examples/custom-rlpx-subprotocol/Cargo.toml
@@ -16,6 +16,5 @@ reth-provider.workspace = true
 reth-primitives.workspace = true
 tokio-stream.workspace = true
 reth-rpc-types.workspace = true
-reth-tracing.workspace = true
 reth.workspace = true
 eyre.workspace = true
